refactor :: extract bookmark logic
