+++ src/main/java/com/dnd/spaced/domain/bookmark/domain/repository/BookmarkCrudRepository.java
@@ -1,6 +1,6 @@
-package com.dnd.spaced.domain.word.domain.repository;
+package com.dnd.spaced.domain.bookmark.domain.repository;
 
-import com.dnd.spaced.domain.word.domain.Bookmark;
+import com.dnd.spaced.domain.bookmark.domain.Bookmark;
 import org.springframework.data.repository.CrudRepository;
 
 interface BookmarkCrudRepository extends CrudRepository<Bookmark, Long> {
