+++ src/main/java/com/dnd/spaced/global/exception/ExceptionCode.java
@@ -34,5 +34,6 @@ public enum ExceptionCode {
     UNAUTHORIZED_COMMENT,
     INVALID_REASON_NAME,
     REPORTED_COMMENT_NOT_FOUND,
-    FORBIDDEN_REPORT_ACCOUNT
+    FORBIDDEN_REPORT_ACCOUNT,
+    BOOKMARK_WORD_NOT_FOUND
 }
